requirements.txt
streamlit==1.37.1
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
python-dateutil==2.9.0.post0
pytz==2024.1
python-dotenv==1.0.1
